Added no images error

+++ commands/images.js
@@ -53,6 +53,7 @@ module.exports = {
                         pages.push(result.url);
                     }
                 });
+                if(!pages.length) return message.edit("Could not find any images");
                 msg.delete();
                 message.channel.send(pages[0]).then(msg => msg.react('⏪').then(() => msg.react('⏩').then(() => {
                     let page = 0;
